# [opentype/gpos] use plain struct instead of pointer