During some benchmarks, I found `tables.PairSet.FindGlyph` was a really high source of heap allocations.
Using a plain struct (probably) allows the compiler/runtime to allocate it on the stack, resulting in a sizable reduction of heap allocations.