@benoitkugler 2023-11-29
> @benoitkugler Do you have before/after benchmark info you could share?

Here you are : the code simply runs Shaper.Shape, with a font using a GPOS table (Arimo).

<details><summary>Code</summary>
<p>

	face := loadOpentypeFont(b, "/usr/share/fonts/truetype/croscore/Arimo-Regular.ttf")
	textInput := []rune("Une jolie phrase en français !")
	input := Input{
		Text:      textInput,
		RunStart:  0,
		RunEnd:    len(textInput),
		Direction: di.DirectionLTR,
		Size:      16 * 72,
		Face:      face,
		Script:    language.Latin,
		Language:  language.NewLanguage("fr"),
	}
	shaper := HarfbuzzShaper{}

	b.ResetTimer()

	for i := 0; i < b.N; i++ {
		_ = shaper.Shape(input)
	}

</p>
</details> 

### Benchmark stats
   
```
       │   old.txt   │              new.txt               │
       │   sec/op    │   sec/op     vs base               │
GPOS-4   40.91µ ± 2%   37.47µ ± 5%  -8.41% (p=0.000 n=10)

       │   old.txt    │               new.txt                │
       │     B/op     │     B/op      vs base                │
GPOS-4   15.94Ki ± 0%   11.88Ki ± 0%  -25.48% (p=0.000 n=10)

       │  old.txt   │              new.txt               │
       │ allocs/op  │ allocs/op   vs base                │
GPOS-4   253.0 ± 0%   227.0 ± 0%  -10.28% (p=0.000 n=10)
``` 

### Memory profile - Before: 
![old](https://github.com/go-text/typesetting/assets/31929422/40a81f72-e1a9-4a57-a093-5217ccc0d5a5)

### Memory profile - After : 
![new](https://github.com/go-text/typesetting/assets/31929422/049c4673-0e55-4e14-ab41-942277b2a925)