@andydotxyz 2023-11-29
Has the "After" trimmed the RunN contribution or did that go to 0?
I can't do the maths accurately on the fly but it looks like memory is just being moved to other places rather than reduced overall.