@benoitkugler 2023-11-29
> Has the "After" trimmed the RunN contribution or did that go to 0? I can't do the maths accurately on the fly but it looks like memory is just being moved to other places rather than reduced overall.

My screenshot was incomplete, see the new one : 

![Capture d’écran du 2023-11-29 12-15-00](https://github.com/go-text/typesetting/assets/31929422/5f71f892-9e19-4292-bafe-d14665a6ac5b)